SFT/requirements.txt
# py310 for aguvis training
accelerate==1.11.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asteval==1.0.6
async-timeout==5.0.1
asyncio==3.4.3
attrs==25.3.0
azure-ai-ml==1.27.1
azure-common==1.1.28
azure-containerregistry==1.0.0
azure-core==1.34.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-data-tables==12.4.0
azure-identity==1.23.0
azure-keyvault==4.2.0
azure-keyvault-certificates==4.9.0
azure-keyvault-keys==4.10.0
azure-keyvault-secrets==4.9.0
azure-mgmt-containerregistry==14.0.0
azure-mgmt-core==1.5.0
azure-mgmt-resource==23.4.0
azure-mgmt-storage==23.0.0
azure-monitor-opentelemetry==1.6.9
azure-monitor-opentelemetry-exporter==1.0.0b36
azure-storage-blob==12.25.1
azure-storage-file-datalake==12.20.0
azure-storage-file-share==12.21.0
backports.tarfile==1.2.0
beautifulsoup4==4.13.4
bitarray==3.4.2
bitstring==4.3.1
blinker==1.9.0
Brotli==1.0.9
cachetools==5.0.0
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.8
click-option-group==0.5.7
click-plugins==1.1.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.2
cycler==0.12.1
datasets==3.6.0
deepspeed==0.16.0
Deprecated==1.2.18
dill==0.3.8
distlib==0.3.9
distro==1.9.0
docker-pycreds==0.4.0
ecdsa==0.19.1
einops==0.8.1
exceptiongroup==1.3.0
filelock==3.13.1
fixedint==0.1.6
Flask==2.3.3
flask-cors==6.0.0
Flask-Executor==1.0.0
Flask-FontAwesome==0.1.5
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.43
gmpy2==2.2.1
google-api-core==2.24.2
google-auth==2.40.2
googleapis-common-protos==1.70.0
h11==0.16.0
hf-xet==1.1.2
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
humanize==4.12.3
identify==2.6.9
idna==3.7
importlib_metadata==8.6.1
intel-cmplr-lib-ur==2025.1.1
intel-openmp==2025.1.1
isodate==0.7.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keyring==25.6.0
kiwisolver==1.4.8
liger_kernel==0.5.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.15.0
matplotlib==3.10.3
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
msgpack==1.1.0
msrest==0.7.1
multidict==6.5.0
multiprocess==0.70.16
networkx==3.4.2
ninja==1.11.1.4
nodeenv==1.9.1
numpy==2.0.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
openai==1.84.0
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.14
opencensus-ext-logging==0.1.1
opencv-python-headless==4.11.0.86
opentelemetry-api==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-dbapi==0.52b1
opentelemetry-instrumentation-django==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-instrumentation-flask==0.52b1
opentelemetry-instrumentation-psycopg2==0.52b1
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-urllib==0.52b1
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-wsgi==0.52b1
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
packaging==24.2
pandas==2.2.3
passpy==1.0.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
pre_commit==4.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.4
psutil==6.1.1
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pydash==8.0.5
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pyperclip==1.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-gnupg==0.5.4
pytz==2025.2
PyYAML==6.0.2
qwen-vl-utils==0.0.14
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rich-click==1.8.9
rpds-py==0.25.1
rsa==4.9.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
SecretStorage==3.3.3
sentry-sdk==2.25.1
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sshpubkeys==3.3.1
strictyaml==1.7.3
sympy==1.13.1
tabulate==0.9.0
tbb==2022.1.0
tcmlib==1.3.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.57.3
triton==3.1.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
umf==0.10.0
urllib3==2.3.0
virtualenv==20.30.0
wandb==0.18.3
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
xdg==6.0.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.21.0
